backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
requests==2.31.0
httpx==0.24.1
pytest==8.3.5
anyio==3.7.1
starlette==0.27.0
numpy
scikit-learn
xgboost
imblearn
joblib
pydantic
sqlalchemy
python-dotenv
